+++ apps/pmb/app/about/page.tsx
@@ -1,40 +1,7 @@
 'use client';
-import React, { FC, ReactElement, useRef, useEffect } from 'react';
-import { Navbar, Footer, Button } from '@uninus/components';
+import { FC, ReactElement } from 'react';
+import { Navbar, Button, Reveal } from '@uninus/components';
 import Image from 'next/image';
-import { motion, useInView, useAnimation } from 'framer-motion';
-
-interface Props {
-  children: JSX.Element;
-  w?: 'fit' | 'full';
-}
-
-const Reveal = ({ children, w = 'fit' }: Props) => {
-  const ref = useRef(null);
-  const isViewed = useInView(ref, { once: true });
-  const animateControllers = useAnimation();
-
-  useEffect(() => {
-    if (isViewed) {
-      animateControllers.start('visible');
-    }
-  }, [isViewed]);
-  return (
-    <div ref={ref} className={`relative ${w} overflow-hidden `}>
-      <motion.div
-        variants={{
-          hidden: { opacity: 0, y: 75 },
-          visible: { opacity: 1, y: 0 },
-        }}
-        initial="hidden"
-        animate={animateControllers}
-        transition={{ duration: 0.5, delay: 0.25 }}
-      >
-        {children}
-      </motion.div>
-    </div>
-  );
-};
 
 const About: FC = (): ReactElement => {
   return (
@@ -241,9 +208,6 @@ const About: FC = (): ReactElement => {
             </div>
           </Reveal>
         </div>
-      </Reveal>
-      <Reveal>
-        <Footer />
       </Reveal>
     </section>
   );
