+++ libs/components/src/atoms/index.ts
@@ -3,3 +3,4 @@ export * from './button/interface';
 export * from './button/enums';
 export * from './icons';
 export * from './loadings';
+export * from './reveal';
