+++ libs/components/src/atoms/reveal/type.ts
@@ -1,0 +1,4 @@
+export type TRevealProps = {
+  children: JSX.Element;
+  w?: 'fit' | 'full';
+};
